feat: add autocompletion, autoformat, and telescope plugins for nvim
