+++ .config/nvim/lua/plugins/oil.lua
@@ -1,0 +1,16 @@
+-- File manager
+return {
+	"stevearc/oil.nvim",
+	---@module 'oil'
+	---@type oil.SetupOpts
+	opts = {
+		view_options = {
+			show_hidden = true,
+		},
+	},
+	-- Optional dependencies
+	dependencies = { "nvim-tree/nvim-web-devicons" },
+	-- Lazy loading is not recommended because it is very tricky to make it work correctly in all situations.
+	lazy = false,
+}
+
